Create useLayerSelect hook
